pyproject.toml
[project]
name = "general-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "crawl4ai>=0.7.3",
    "ollama>=0.6.1",
    "openai>=1.99.9",
    "opencode-ai>=0.1.0a36",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]
